+++ pgdrive/scene_creator/road/road_network.py
@@ -3,7 +3,7 @@ import logging
 from typing import List, Tuple, Dict
 import numpy as np
 from pgdrive.utils.scene_utils import get_road_bounding_box
-from pgdrive.utils.scene_utils import Decoration
+from pgdrive.utils.constans import Decoration
 from pgdrive.scene_creator.lanes.lane import LineType, AbstractLane
 from pgdrive.scene_creator.lanes.straight_lane import StraightLane
 from pgdrive.scene_creator.road.road import Road
