+++ pgdrive/utils/constans.py
@@ -1,0 +1,15 @@
+class Decoration:
+    start = "decoration"
+    end = "decoration_"
+
+
+class Goal:
+    """
+    Goal at intersection
+    The keywords 0, 1, 2 should be reserved, and only be used in roundabout and intersection
+    """
+
+    RIGHT = 0
+    STRAIGHT = 1
+    LEFT = 2
+    ADVERSE = 3  # Useless now
